### 2019-08-17 #1654 @henrikmidtiby
> PDF failed to compile for issue #1654 with the following error:
This commit should fix the issue with the missing orcid id for Henrik Dyrberg Egemose.
https://github.com/henrikmidtiby/PorpoiseTracker/commit/c3bb5f7eec5936373d553306e56ed58eeeabdc5c
