### 2019-08-18 #1654 @henrikmidtiby
søn. 18. aug. 2019 kl. 19.26 skrev ooo[bot] <notifications@github.com>:

> 👋 Hey @arfon <https://github.com/arfon>...
>
> Letting you know, @kthyng is currently OOO until Thursday, August 22nd
> 2019
> <https://github.com/openjournals/joss-reviews/issues/1619#issuecomment-518038270>.
> ❤️
>
> —
>
>
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/openjournals/joss-reviews/issues/1654?email_source=notifications&email_token=AAALKBIBBMZXPXQONHJ7LY3QFGA5ZA5CNFSM4IMK2IKKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD4REMWI#issuecomment-522339929>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAALKBNZRTXXKDMEE52NPMDQFGA5ZANCNFSM4IMK2IKA>
> .
>
