### 2019-08-19 #1654 @henrikmidtiby
The article proof look fine.
