### 2019-08-26 #1654 @jizecn
> > @kthyng Can you change the repository to https://github.com/henrikmidtiby/PorpoiseTracker ?
> > Right now the repository points towards a url, which only contains the paper and not the source code. If it is difficult to change, I can resubmit the paper with the correct information.
>
> I've made this update.

The embedded link in the sentence of
_Repository: Is the source code for this software available at the repository url?_
is actually still pointing to the wrong url.
