### 2019-08-26 #1654 @jizecn
Thanks for suggesting me to review the code/paper submission @henrikmidtiby

I have done a quick read of the paper, but have not managed to proceed to test the code as instructed (due to different operating system I use (Mac OS).).

Correct me if I am wrong, since I have not looked into the code in detail. I think the paper subject is a little misleading. Porpoise Tracker, to me, implies automatic detection and track of Porpoise or similar, where the work provides a GUI based application that allows importing Drone logs and video footages and the measure of Porpoise sizes is a manual process. The work seems to me aiming for end users of Porpoise behaviour researchers. Therefore, I suggest the authors to make the target audiences more explicit in the statement. Otherwise, you may attract computer vision scientists seeking target tracking algorithms solutions.

The paper states that "The accuracy of method was investigated by [@Midtiby2019Havforsker]". However, I could not find the link to the reference cited. If possible, can you add a DOI or link to the reference, if it can be made publicly available online.

Other minor remarks:
* There is a small typo in the document, "To to that you can..." should be "To do that ...".
* You might want to add a short description of the project on your github repo.

This is my first review for JOSS. I hope the above makes sense. Thanks.
