### 2019-08-29 #1654 @kthyng
@henrikmidtiby Have you or any colleagues installed this software on a Mac? Or do you have access to one to try? Ideally, I would hope between you and @jizecn we can get this software installed on a mac, or there could be an alternative setup like a virtual machine or docker image so that @jizecn can review your submission. Expanding your installation possibilities would also improve your submission.
