### 2019-09-01 #1654 @henrikmidtiby
@kthyng We attempted to install it on a mac six months ago, but it failed due to some missing requirements / dependencies. We haven't tried to install it on MacOS since.
I will test whether the program installs in a virtual machine and report back.
