### 2019-09-02 #1654 @henrikmidtiby
@kthyng I have found some unexpected issues related to graphics drivers and the oracle virtual box emulator that I have set up. This hinders the porpoise tracker to load a video and makes rest of the program useless. I am searching for a workaround.
