### 2019-09-09 #1654 @kthyng
I'll just type a ping here for everyone. @henrikmidtiby were you able to make any progress on a workaround, or are you still in the process?
