### 2019-09-20 #1654 @henrikmidtiby
@kthyng I got stuck in the workaround and was distracted due to real life events afterwards. My hope to get it to work is quite small at the moment, as the error messages that I battle with does not help me identify what is wrong with my virtual machine.
