### 2019-10-11 #1654 @abhishekbajpayee
I tried installing. Seems like `libcairo2-dev` needs to be added to the `sudoapt install ...` list. Even after installing `libcairo2-dev`, when I run `pip install -r requirements.txt`, I see long error because of:
```
No package 'py3cairo' found
```
Also, it seems if someone installs from a fresh clean OS, `apt install python3-pip` will be needed. FYI, I am installing in a fresh `ubuntu:18.04` docker container.

@henrikmidtiby could you address some of these issues?
