### 2019-10-15 13b76ac
Updated installation guide.
