### 2019-10-15 #1654 @henrikmidtiby
@abhishekbajpayee I have pushed an updated installation guide to github.
https://github.com/henrikmidtiby/PorpoiseTracker/commit/13b76acc2c7cdbf36ab824512cd7d16c24f631db

The installation guide was tested on a minimal installation of Ubuntu 18.04.3 LTS, where the first thing I did after updating all installed packages was to install the porpoise tracker program.
