### 2019-10-15 #1654 @henrikmidtiby
@abhishekbajpayee I have tried to install the porpoise tracker in a docker container, but is stuck on the following issue.

```
Unable to init server: Could not connect: Connection refused
Unable to init server: Could not connect: Connection refused
```

The approach is based on the Dockerfile shown below. Keep in mind that this is my first attempt at using docker ...
```
FROM ubuntu:18.04
LABEL maintainer="Henrik Skov Midtiby"

ENV SHELL /bin/bash

RUN apt-get update && \
    apt-get install -y python3-venv wine64 gstreamer1.0-libav gobject-introspection python3-gi python3-gi-cairo gir1.2-gtk-3.0 libgirepository1.0-dev git python-gst-1.0 ffmpeg libcairo2-dev pkg-config python3-dev

ENV VIRTUAL_ENV=/opt/venv
RUN python3 -m venv $VIRTUAL_ENV
ENV PATH="$VIRTUAL_ENV/bin:$PATH"

RUN mkdir -p /usr/local/src/porpoisetracker
RUN cd /usr/local/src/porpoisetracker && \
	git clone https://github.com/henrikmidtiby/PorpoiseTracker.git
RUN cd /usr/local/src/porpoisetracker/PorpoiseTracker && \
	pip install -r requirements.txt
RUN cd /usr/local/src/porpoisetracker && \
	git clone https://github.com/henrikmidtiby/gtk_modules.git
RUN cd /usr/local/src/porpoisetracker/gtk_modules && \
	pip install -e .

WORKDIR /usr/local/src/porpoisetracker
```
