### 2019-11-15 #1654 @kthyng
Hi @henrikmidtiby. This review has stalled quite a bit due to the inability of the reviewers to run your code. Could you find a local colleague and/or collaborator to help you either generalize the installation process or understand the issues that the reviewers are having so they can move past them? There is a limit to the amount of effort we can reasonably ask of a reviewer in this part of the process. I propose that I pause this review in the meantime. Does this sound reasonable to you?
