### 2019-11-15 #1654 @abhishekbajpayee
Here is update from my end. I tried following install instructions again and was able to get the installation working using the Dockerfile that @henrikmidtiby pasted above. But using this container, I am not able to actually run the GUI application.

Part of the issue is that I am reluctant to `sudo apt install` things on my personal machine for a review. This is why using a container is best and keeps everything isolated. So it would be ideal to have install and run instructions for a container.

@kthyng what are your thoughts on this? Let me know if my expectation is unreasonable. If that is the case, I can just modify my local environment for this review and move this forward. However, it would be difficult for me to do more reviews (which I would like to help with) if I have to modify the environment i.e. continue `sudo apt install`ing things on my personal machine every time.
