### 2019-11-15 #1654 @henrikmidtiby
@kthyng I support your suggestion to pause the submission, until the installation procedure is explained in enough detail. How should I proceed from here?

I understand the reluctance by @abhishekbajpayee to install packages in the global scope.
It would be much better if we could get the docker image or something similar to work.
