### 2020-02-05 #1654 @egemose
Hi @kthyng I have made a new submission of a rewritten version of this software to ease the installation and use, You are also the Editor at the new submission over at https://github.com/openjournals/joss-reviews/issues/2056
